Fix formatting of Binary Operations
The formatting of the Binary Operations portion of the Operations in Class Set was messed up (due to the pipe character being interpreted as part of the table specification). Unfortunately, there seems to be no way to escape a pipe character within backticks, so I've had to resort to using <code> tags.