replaces dynamic content generation with static pages
Unfortunately migration to jekyll 1.0.2 has broken our website.
Therefore I've built it locally with jekyll 0.12.1 and then replaced
all the dynamic content by the contents of the _site directory produced
by jekyll. The finishing touch was to add a `.nojekyll` file to the root
of the repo to disable Jekyll.