Moved to standard notation in performance characteristics docs
Notation used is now the standard computer-science one; a.k.a.: "Big-O" Notation from the Bachmann-Landau family.

Also allowed me to remove the legend table; as the notation is standard. Additionally added in `<sup>` for the more specific amoritized complexities and the like. (often represented with `*` notation, but this'll suffice)