Fix title capitalization
The language backing Liquid templates is quite dubios and the `capitalize`
filter wasn't actually doing what was intended. As such, a `partof` property
with a value of "foo-bar" was rendered as ['foo', 'bar'] in the title.

Using a plugin seems to be the easiest way to solve this problem.

Solution from here: http://stackoverflow.com/a/23453152/58808