Final improvement in `accepted` clause
After some thought, I believe that the compiler maintainers have a
better judgement of when proposals can be merged, especially because of
the compatibility guarantees that the Scala compiler contract enforces.

In this way, I propose that the Compiler team chooses the most
appropriate version to include the accepted proposal.