Added an example on how to use anonymous functions in the "Anonymous … (#628)
Added an example on how to use anonymous functions in the "Anonymous Function Syntax" page (tutorials/tour/anonymous-function-syntax.md).

Omitted the, apparently for Scala unconventional, usage of 'var's in the code.
Example is now slimmer, i.e. there is only one line in the main method where the anonymous function is called once and the return value is immediately printed.