style guide: revise sections on infix and postfix notation
the postfix text is mostly the same, but now uses the word "postfix"
instead of "suffix"

the infix text is totally rewritten and now more strongly discourages
use of infix for alphabetic method names.  back when the style guide
was first written, I think there was more experimentation with, and
more acceptance of, such use of infix notation.  these days, not so
much.

I omitted the material about mixing dot and dotless notation -- the "a
b.(c) d" example.  I found that passage highly confusing and I've seen
others be confused by it as well.