Overhaul the previous specification
According to feedback from the SIP Committee, this commit adds changes
to the previous specification to fit type aliases, it describes the
technical details and some implementation details that I have dealt with
in my work-in-progress implementation.

The proposal is not yet ready, though. The motivation section is
missing, for example.