Add "build project" step and misc. improvements (#1055)
* indent code blocks to improve style

This indents the code blocks associated with a given list item so that
they are continuations of the list items.

* fix list ordering to improve style

This converts this list item to a ordered item so that the ordered list
is not broken by the unordered list item.

* group list items related to ScalaTest dependency

This groups the list items related to adding the ScalaTest dependency to
improve readability.

* grammar change

Rather than personifying the document or referring to the author(s) just
state that the tutorial assumes prior knowledge.

* rewrap a few lines

This rewraps a few lines to improve readability.

* add "build project" step

For me, after `sbt` downloads ScalaTest the `CubeCalculatorTest` class
will still have several errors because ScalaTest has not been
compiled/indexed by IntelliJ IDEA so this explicitly adds a
"build project" step to trigger that.

* correctly identify cause of FunSuite and test() not being recognized

When I followed this tutorial I either did it really fast or in the
wrong order and ended up with `FunSuite` and `test()` not being
recognized.  I mistakenly thought building the project fixed it but it
turns out I just didn't notice that `sbt` hadn't finished syncing.