Clarifications for List blocks
Having just spent a couple of hours trying to figure out how to make list blocks work, here's a tweak to provide the information that would have made my life easier.  It was previously taking for granted a few details that may be obvious if you are experienced in specific wikitext languages, but aren't otherwise.

(In particular, the fact that you *must* have extra space before the bullets, that you may *only* use, eg, `1.` as a lead-in identifier, not `2.`, and that additional spaces create sub-levels.  None of those are true of [standard Markdown](https://daringfireball.net/projects/markdown/syntax), which is the most common such language.)