Update files.md
I propose to remove this part, because it's not a real recommendation. Something in the style "do whatever you want if it is good for you". I think that putting multiple entities in a single file is certainly a bad practice. Especially when we speak about a large complex project.

"Another case is when multiple classes logically form a single, cohesive
group, sharing concepts to the point where maintenance is greatly served
by containing them within a single file. These situations are harder to
predict than the aforementioned sealed supertype exception. Generally
speaking, if it is *easier* to perform long-term maintenance and
development on several units in a single file rather than spread across
multiple, then such an organizational strategy should be preferred for
these classes. However, keep in mind that when multiple units are
contained within a single file, it is often more difficult to find
specific units when it comes time to make changes."