Updated the comment for the return type
The previous inline code highlighting of `: Int` together was confusing: the sentence as a whole seemed to imply that the return type could be found _after_ the colon *and* the `Int`, i.e. the sentence could be read to imply that `: Int` was a separate syntactic construct from the return type, even though the return type `Int` was already there.