Updates Library Author Guide regarding Nexus
As far as I can tell, the current description does not correspond to the correct place to look anymore. There is still a Staging repository in "User Managed Repository" (the type selected by default), but it does not contain a library waiting to be released.
Since it was a head-scratch for me when publishing my first lib, I assume it might help others.