Change style of h4 to pop less out
The h4 style dominated h3 which was problematic for larger
bodies of text. This PR proposes a slightly smaller font-size
and not using bold-face. It also adjusts the margins slightly
to be closer to the textbody.

The style is only rarely used and the effect can be observed
on the following page:

   /overviews/core/actors.html