reboot the Scala FAQ
gah, nobody has touched this thing in years. and, perhaps, because
it's so neglected, nobody in our chat rooms and forums ever refers
anyone here.

let's do something about it.  changes so far:

* don't duplicate content from Stack Overflow; this is a worthy
  but IMO failed experiment
* discourage long writeups; encourage links to answers elsewhere
* add a few general questions at the top that I know to be
  especially frequently asked
* drop the question about `breakOut`, it is not an FAQ these days

this PR is just a start, much room for improvement remains,
but IMO this is already mergeable