Drop the links to Scala 2 and Scala 3 pages and integrate on landing page
The links were not visible enough and made it difficult to discover the additional
content. Instead all content is now presented on the landing page and can simply
be discovered by scrolling. This made the Scala 2 page obsolete (at least for now).

I keep the Scala 3 page since we want to be able to link to the "Scala 3" documentation.