Fix multiple highlighted headlines in TOC
Previously, multiple equally named headlines were highlighted as the
current page, if the name was the same as on the actually current page.
This issued showed up in the book, where for instance "Summary" was used
more than once. Instead, here I am using the page ordinal instead of the
title.

It should be safe to assume that the page has a number, since otherwise
it would not show up in the TOC, IIUC.