A Tutorial on writing Scala apps on Android
This tutorial is a shortened and updated version of the tutorial from my blog note:
https://makingthematrix.wordpress.com/2021/04/07/how-to-build-an-android-app-in-scala-2-13/

I focus on a basic example that uses GraalVM Native Image to create an Android app, but in the introduction
I mention that there are also other options and at the end I provide links to more details.

The link to the tutorial was added to the popup menu just under "Learn" -> "Scala for Java Programmers".