Tutorials as a new category on the overviews page
Currently some of the learning materials are accessed through the "overviews" page while others are under the "Learn" meanu in the header.
The header is not intuitive. It's not visible on the main page, so the user needs to click on the overviews to be able to access the materials
under the header, which defeats the purpose of having them there if there ever was one. (I suspect the current situation is a side-effect of
some other changes, not a conscious decision). On top of that, "Scala with Maven" and "Binary Compatibility for Library Authors" are nowhere
to be found.

My proposal is to remove the "Learn" menu from the header and instead to have a new category in the "overviews" page, called "Tutorials",
directly under "Language", so it's pretty high on the webpage and easy to find. I discussed it on the Scala discord and I got an impression
that even though it's not a perfect solution, it's at least a step in the right direction.

I moved all the links from the "Learn" menu to that new category and added short descriptions and authors' names in places where I was pretty
sure what to do. Please review also from this angle: maybe there should more names under the author's tag or it's better to remove them
completely, and maybe someone could update the descriptions.