Fix the Docker setup
I found that our previous Docker setup was not good enough and was failing on some machines.

I propose the following changes, based on what has been done on the scala.epfl.ch website. The main “fix” is that we indicate a specific version of Ruby, rouge, and bundler. We also propose a different way to invoke docker-compose up that writes the generated files as a regular user (instead of root).