Update implicit-conversions.md (#2674)
* Update implicit-conversions.md

I'm not sure if it is intentional, but the Tour-of-Scala documentation does not mention how to bring implicit conversions into scope. In fact, that information is a bit hard to find anywhere else, so I thought I'll propose to add it here?

* Update _tour/implicit-conversions.md from suggestion

Co-authored-by: Jamie Thompson <bishbashboshjt@gmail.com>

* Update implicit-conversions.md

Improved Scala 3 description. Thanks for all the suggestions!

* Fix import in _tour/implicit-conversions.md

Co-authored-by: Jamie Thompson <bishbashboshjt@gmail.com>

* describe Scala 2 implicit conversion in companion object

* Remove mdoc from Scala 3 example

* Update _tour/implicit-conversions.md

* Update _tour/implicit-conversions.md

* Remove mdoc from Scala 2 example as well

Sorry for wasting your time, next time, I'll install the toolchain.

* merge section "How are implicit conversions selected?"

Co-authored-by: Jamie Thompson <bishbashboshjt@gmail.com>