repaired links to now dead `scalamacros.org`;
the website `scalamacros.org` is no longer being used for what it was
originally intended for when this documentation was written, so I have
linked to the github repository which contains the original data.

It may be worth it to look into `archive.org` for this purpose, if it is
felt that the repository used is not sufficient.
In the event the repository disapears, I have
created a fork, and plan use it for archival purposes. This fork is
welcome to use by the scala project in such an event.
[https://github.com/bicknrown/scalamacros.github.com/](https://github.com/bicknrown/scalamacros.github.com/)