Update extractor-objects.md
The customerid match case does not run in Scastie for Scala 3 without curly braces.

I've added them here.

Is this a problem with Scastie's Scala 3? If so, you can reject the PR and fix Scastie.

Is this a fix for the tutorial code for Scala 3?  If so, please accept this PR.