string-interpolation: the raw-interpolator also performs variable substitution
Clarify that the raw-interpolator also performs variable
substitution. This was not obvious from the text and it is easy to
believe that raw means just that: the raw string, as is, without any
substitutions performed.

This also moves the last sentence of the raw-interpolator section
into the following section, where it logicaly belongs.