Remove incorrect reference to "Java parameter name length"
There is no practice in the Java community or "ecosystem" that would discourage short parameter names. On the contrary even the Java standard libraries often use single letter variable names when the purpose of the parameter is evident from the context.

So I have removed the statement as it is simply incorrect.